api/requirements.txt
asgiref==3.8.1
attrs==24.2.0
coverage==7.2.7
django-cors-headers==4.3.0
django-environ==0.11.2
django-nose==1.4.7
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.27.2
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
jsonschema-specifications==2023.12.1
jsonschema==4.23.0
packaging==24.1
Pillow==10.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
python-dotenv==0.21.0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1